+++ src/intersperse.ts
@@ -1,10 +1,14 @@
-import { getIterator } from "./getIterator.js";
+import type { AsynchronousIterable } from "@vangware/types";
+import { flat } from "./flat.js";
+import { initial } from "./initial.js";
+import { repeat } from "./repeat.js";
+import { zip } from "./zip.js";
 
 /**
- * Add the given `separator` between each element of the given `iterable`.
+ * Add the given `separator` between each element of the given iterable or
+ * asynchronous iterable.
  *
  * @category Generators
- * @category Mappers
  * @example
  * ```typescript
  * const intersperseComma = intersperse(",");
@@ -13,23 +17,9 @@ import { getIterator } from "./getIterator.js";
  * @param separator Separator to add between each element.
  * @returns Curried function with `separator` in context.
  */
-export const intersperse = <Separator>(separator: Separator) =>
-	function* <Item>(iterable: Iterable<Item>) {
-		const iterator = getIterator(iterable);
-		const item = { done: false, ...iterator.next() };
+export const intersperse = <Separator>(separator: Separator) => {
+	const repeatSeparator = repeat(separator)(Infinity);
 
-		// eslint-disable-next-line functional/no-loop-statement
-		while (!item.done) {
-			const next = { done: false, ...iterator.next() };
-
-			yield item.value;
-
-			// eslint-disable-next-line functional/no-conditional-statement
-			if (!next.done) {
-				yield separator;
-			}
-
-			// eslint-disable-next-line functional/immutable-data, functional/no-expression-statement
-			Object.assign(item, next);
-		}
-	};
+	return <Item>(iterable: AsynchronousIterable<Item>) =>
+		initial(flat(zip(iterable)(repeatSeparator)));
+};