+++ src/some.ts
@@ -1,11 +1,13 @@
-import { isArray } from "@vangware/predicates";
-import type { Predicate } from "@vangware/types";
+import { isBoolean } from "@vangware/predicates";
+import type { AsynchronousIterable, Predicate } from "@vangware/types";
+import { every } from "./every.js";
+import type { ReducerOutput } from "./types/ReducerOutput.js";
 
 /**
- * Evaluates items in an iterable against a predicate and returns `true` if any
- * item evaluates to `true`.
+ * Evaluates items in an iterable or asynchronous iterable against a predicate
+ * and returns `true` if any item evaluates to `true`.
  *
- * @category Validators
+ * @category Reducers
  * @example
  * ```
  * const someEven = some((number: number) => number % 2 === 0);
@@ -15,21 +17,20 @@ import type { Predicate } from "@vangware/types";
  * @param predicate Predicate function to evaluate each item.
  * @returns Curried function with `predicate` set in context.
  */
-export const some =
-	<Item, Predicated extends Item>(predicate: Predicate<Item, Predicated>) =>
-	(iterable: Iterable<Item>) => {
-		// eslint-disable-next-line functional/no-conditional-statement
-		if (isArray(iterable)) {
-			return iterable.some(item => predicate(item));
-		} else {
-			// eslint-disable-next-line functional/no-loop-statement
-			for (const item of iterable) {
-				// eslint-disable-next-line functional/no-conditional-statement
-				if (predicate(item)) {
-					return true;
-				}
-			}
+export const some = <Item, Predicated extends Item = Item>(
+	predicate: Predicate<Item, Predicated>,
+) => {
+	const everyPredicate = every<Item>(item => !predicate(item));
 
-			return false;
-		}
+	return <Iterable extends AsynchronousIterable<Item>>(
+		iterable: Iterable,
+	) => {
+		const output = everyPredicate(iterable);
+
+		return (
+			isBoolean(output)
+				? !(output as unknown as boolean)
+				: output.then(result => !result)
+		) as ReducerOutput<Iterable, boolean>;
 	};
+};