+++ tests/entriesToObject.test.ts
@@ -1,18 +1,18 @@
 import type { Tests } from "@vangware/test";
 import type { ReadOnlyRecord } from "@vangware/types";
 import { entriesToObject } from "../src/entriesToObject.js";
-import { iterateArray } from "./iterateArray.js";
+import { asyncIterateArray, iterateArray } from "./utils.js";
 
-const stringKey = "key";
-const numberKey = 1;
+const stringKey = "key" as const;
+const numberKey = 1 as const;
 const symbolKey = Symbol("key");
-const value = "value";
+const value = "value" as const;
 
 export default [
 	{
 		given: "an array of entries [string, value]",
 		must: "return an object with the shape { string: value }",
-		received: entriesToObject([[stringKey, value]]),
+		received: entriesToObject([[stringKey, value] as const]),
 		wanted: { [stringKey]: value },
 	},
 	{
@@ -30,19 +30,43 @@ export default [
 	{
 		given: "an iterable of entries [string, value]",
 		must: "return an object with the shape { string: value }",
-		received: entriesToObject(iterateArray([[stringKey, value]])),
+		received: entriesToObject(iterateArray([[stringKey, value] as const])),
 		wanted: { [stringKey]: value },
 	},
 	{
 		given: "an iterable of entries [number, value]",
 		must: "return an object with the shape { number: value }",
-		received: entriesToObject(iterateArray([[numberKey, value]])),
+		received: entriesToObject(iterateArray([[numberKey, value] as const])),
 		wanted: { [numberKey]: value },
 	},
 	{
 		given: "an iterable of entries [symbol, value]",
 		must: "return an object with the shape { symbol: value }",
-		received: entriesToObject(iterateArray([[symbolKey, value]])),
+		received: entriesToObject(iterateArray([[symbolKey, value] as const])),
+		wanted: { [symbolKey]: value },
+	},
+	{
+		given: "an async iterable of entries [string, value]",
+		must: "return an object with the shape { string: value }",
+		received: entriesToObject(
+			asyncIterateArray([[stringKey, value] as const]),
+		),
+		wanted: { [stringKey]: value },
+	},
+	{
+		given: "an async iterable of entries [number, value]",
+		must: "return an object with the shape { number: value }",
+		received: entriesToObject(
+			asyncIterateArray([[numberKey, value] as const]),
+		),
+		wanted: { [numberKey]: value },
+	},
+	{
+		given: "an async iterable of entries [symbol, value]",
+		must: "return an object with the shape { symbol: value }",
+		received: entriesToObject(
+			asyncIterateArray([[symbolKey, value] as const]),
+		),
 		wanted: { [symbolKey]: value },
 	},
 ] as Tests<ReadOnlyRecord<string, string>>;