+++ tests/every.test.ts
@@ -1,6 +1,6 @@
 import type { Tests } from "@vangware/test";
 import { every } from "../src/every.js";
-import { iterateArray } from "./iterateArray.js";
+import { asyncIterateArray, iterateArray } from "./utils.js";
 
 const everyNumbers = every(
 	(value: unknown): value is number => typeof value === "number",
@@ -31,6 +31,18 @@ export default [
 		given: "an iterable of numbers with a string on it",
 		must: "return false",
 		received: everyNumbers(iterateArray(numbersWithStringArray)),
+		wanted: false,
+	},
+	{
+		given: "an async iterable of numbers",
+		must: "return true",
+		received: everyNumbers(asyncIterateArray([0, 1, 2, 3])),
+		wanted: true,
+	},
+	{
+		given: "an async iterable of numbers with a string on it",
+		must: "return false",
+		received: everyNumbers(asyncIterateArray([0, 1, 2, "foo", 3])),
 		wanted: false,
 	},
 ] as Tests<boolean>;