+++ tests/find.test.ts
@@ -1,6 +1,7 @@
 import type { Tests } from "@vangware/test";
+import type { Maybe } from "@vangware/types";
 import { find } from "../src/find.js";
-import { iterateArray } from "./iterateArray.js";
+import { asyncIterateArray, iterateArray } from "./utils.js";
 
 const findString = find(
 	(value: unknown): value is string => typeof value === "string",
@@ -31,4 +32,16 @@ export default [
 		received: findString(iterateArray([0, 1, 2])),
 		wanted: undefined,
 	},
-] as Tests<string | undefined>;
+	{
+		given: "an async iterable of numbers and strings containing search matching item",
+		must: "get the first string that matches",
+		received: findString(asyncIterateArray([0, 1, "foo", 2, "bar"])),
+		wanted: "foo",
+	},
+	{
+		given: "an async iterable of numbers without search matching item",
+		must: "get the first string",
+		received: findString(asyncIterateArray([0, 1, 2])),
+		wanted: undefined,
+	},
+] as Tests<Maybe<string>>;