+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 -   This project adheres to [Semantic Versioning][semver].
 -   This project uses [Gitmoji][gitmoji]
 
+## 2.0.0
+
+-   :arrow_up: dependency upgrade.
+-   :fire: remove all asynchronous utils (the regular utils handle the async
+    cases as well now).
+-   :memo: update docs.
+-   :building_construction: Functions now return an IteratorIterable so it can
+    be reused.
+-   :building_construction: Some reuse and perf optimizations.
+
 ## 1.0.3
 
 -   :arrow_up: dependency upgrade.